+++ backend/src/Routify.Data/Models/TextLog.cs
@@ -22,8 +22,8 @@ public class TextLog
     
     public int InputTokens { get; set; }
     public int OutputTokens { get; set; }
-    public double InputCost { get; set; }
-    public double OutputCost { get; set; }
+    public decimal InputCost { get; set; }
+    public decimal OutputCost { get; set; }
     
     public DateTime StartedAt { get; set; }
     public DateTime EndedAt { get; set; }
