+++ backend/src/providers/Routify.Provider.Core/Completion/CompletionResponse.cs
@@ -6,6 +6,6 @@ public record CompletionResponse
     public CompletionPayload? Payload { get; set; }
     public int InputTokens { get; set; }
     public int OutputTokens { get; set; }
-    public double InputCost { get; set; }
-    public double OutputCost { get; set; }
+    public decimal InputCost { get; set; }
+    public decimal OutputCost { get; set; }
 }
